clean up docker

+++ dkr/create-disk-vol.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
 
-docker container prune -f
+if [ -f auto-prune ]; then docker container prune -f; fi
 docker volume rm -f disk-data
 docker volume create disk-data
 docker run --rm \
