+++ dkr/create-pg-vol.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
 
-docker container prune -f
+if [ -f auto-prune ]; then docker container prune -f; fi
 docker volume rm -f pg-data
 docker volume create pg-data
 docker volume ls
