+++ dkr/run-local.sh
@@ -2,6 +2,5 @@
 
 rm docker-compose.yml
 ln -s docker-compose-local.yml docker-compose.yml
-docker container prune -f
+if [ -f auto-prune ]; then docker container prune -f; fi
 docker-compose up
-
